+++ src/calendar.js
@@ -5821,6 +5821,12 @@ function calendarJs( elementOrId, options, searchOptions ) {
         _element_EventEditorDialog_Title = createElement( "input", null, "text" );
         inputTitleContainer.appendChild( _element_EventEditorDialog_Title );
 
+        _element_EventEditorDialog_Title.onkeydown = function( e ) {
+            if ( e.keyCode === _keyCodes.enter ) {
+                eventDialogEvent_OK();
+            }
+        };
+
         if ( _options.maximumEventTitleLength > 0 ) {
             _element_EventEditorDialog_Title.maxLength = _options.maximumEventTitleLength ;
         }